+++ src/main/java/vn/com/shop/dto/product/ProductVariantResponseDTO.java
@@ -4,13 +4,15 @@ import lombok.Getter;
 import lombok.Setter;
 import vn.com.shop.dto.BaseResponseDTO;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
 public class ProductVariantResponseDTO extends BaseResponseDTO {
     private String sku;
-    private Double price;
-    private Double discountPrice;
+    private BigDecimal price;
+    private BigDecimal discountPrice;
     private Integer stock;
     private Integer reservedStock;
     private String size;
