+++ src/main/java/vn/com/shop/entity/ProductVariantEntity.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Entity
@@ -13,22 +14,14 @@ import java.util.Set;
 public class ProductVariantEntity extends BaseEntity{
     @Column(unique = true)
     private String sku;
-
-    private Double price;
-
-    private Double discountPrice;
-
+    private BigDecimal price;
+    private BigDecimal discountPrice;
     private Integer stock;
-
     private Integer reservedStock;
-
     private String size;
-
     private String color;
-
     @ManyToOne
     private ProductEntity product;
-
     @OneToMany(mappedBy = "variant")
     private Set<InventoryLogEntity> inventoryLogs;
 }
